final commit

+++ UniversiteDomain/DataAdapters/INoteRepository.cs
@@ -4,5 +4,5 @@ namespace UniversiteDomain.DataAdapters;
 
 public interface INoteRepository : IRepository<Note>
 {
-    Task<Note> AddNoteAsync(long etudiantId, long ueId, double valeur);
+    Task<Note> AddNoteAsync(long etudiantId, long ueId, float valeur);
 }
